added event system

+++ engine/Constants.h
@@ -6,8 +6,8 @@
 const int WINDOW_WIDTH{ 1440 };
 const int WINDOW_HEIGHT{ 900 };
 
-const int FPS{ 60 };
-const float FRAME_LENGTH{ 1000 / FPS };
+const float FPS{ 60.f };
+const float FRAME_LENGTH{ 1000.f / FPS };
 
 const int SHIP_SIZE{ 32 };
